--- a/kimchi/src/curve.rs
+++ b/kimchi/src/curve.rs
@@ -1,14 +1,18 @@
 use ark_ec::{short_weierstrass_jacobian::GroupAffine, ModelParameters};
 use commitment_dlog::{commitment::CommitmentCurve, srs::endos};
-use mina_curves::pasta::curves::{
-    pallas::{LegacyPallasParameters, PallasParameters},
-    vesta::{LegacyVestaParameters, VestaParameters},
+use mina_curves::pasta::{
+    curves::{
+        pallas::{LegacyPallas, PallasParameters},
+        vesta::{LegacyVesta, VestaParameters},
+    },
+    Pallas, Vesta,
 };
 use mina_poseidon::poseidon::ArithmeticSpongeParams;
+use o1_utils::fast_msm::msm::MultiScalarMultiplication;
 use once_cell::sync::Lazy;
 
 ///Represents additional information that a curve needs in order to be used with Kimchi
-pub trait KimchiCurve: CommitmentCurve {
+pub trait KimchiCurve: CommitmentCurve + MultiScalarMultiplication {
     /// The other curve that forms the cycle used for recursion
     type OtherCurve: KimchiCurve<
         ScalarField = Self::BaseField,
@@ -25,8 +29,8 @@ pub trait KimchiCurve: CommitmentCurve {
     fn endos() -> &'static (Self::BaseField, Self::ScalarField);
 }
 
-impl KimchiCurve for GroupAffine<VestaParameters> {
-    type OtherCurve = GroupAffine<PallasParameters>;
+impl KimchiCurve for Vesta {
+    type OtherCurve = Pallas;
 
     fn sponge_params() -> &'static ArithmeticSpongeParams<Self::ScalarField> {
         mina_poseidon::pasta::fp_kimchi::static_params()
@@ -36,13 +40,13 @@ impl KimchiCurve for GroupAffine<VestaParameters> {
         static VESTA_ENDOS: Lazy<(
             <VestaParameters as ModelParameters>::BaseField,
             <VestaParameters as ModelParameters>::ScalarField,
-        )> = Lazy::new(endos::<GroupAffine<VestaParameters>>);
+        )> = Lazy::new(endos::<Vesta>);
         &VESTA_ENDOS
     }
 }
 
-impl KimchiCurve for GroupAffine<PallasParameters> {
-    type OtherCurve = GroupAffine<VestaParameters>;
+impl KimchiCurve for Pallas {
+    type OtherCurve = Vesta;
 
     fn sponge_params() -> &'static ArithmeticSpongeParams<Self::ScalarField> {
         mina_poseidon::pasta::fq_kimchi::static_params()
@@ -52,7 +56,7 @@ impl KimchiCurve for GroupAffine<PallasParameters> {
         static PALLAS_ENDOS: Lazy<(
             <PallasParameters as ModelParameters>::BaseField,
             <PallasParameters as ModelParameters>::ScalarField,
-        )> = Lazy::new(endos::<GroupAffine<PallasParameters>>);
+        )> = Lazy::new(endos::<Pallas>);
         &PALLAS_ENDOS
     }
 }
@@ -61,8 +65,8 @@ impl KimchiCurve for GroupAffine<PallasParameters> {
 // legacy curves
 //
 
-impl KimchiCurve for GroupAffine<LegacyVestaParameters> {
-    type OtherCurve = GroupAffine<LegacyPallasParameters>;
+impl KimchiCurve for LegacyVesta {
+    type OtherCurve = LegacyPallas;
 
     fn sponge_params() -> &'static ArithmeticSpongeParams<Self::ScalarField> {
         mina_poseidon::pasta::fp_legacy::static_params()
@@ -72,8 +76,9 @@ impl KimchiCurve for GroupAffine<LegacyVestaParameters> {
         GroupAffine::<VestaParameters>::endos()
     }
 }
-impl KimchiCurve for GroupAffine<LegacyPallasParameters> {
-    type OtherCurve = GroupAffine<LegacyVestaParameters>;
+
+impl KimchiCurve for LegacyPallas {
+    type OtherCurve = LegacyVesta;
 
     fn sponge_params() -> &'static ArithmeticSpongeParams<Self::ScalarField> {
         mina_poseidon::pasta::fq_legacy::static_params()