This PR adds a GPU-accelerated implementation for MSM: https://github.com/supranational/pasta-msm/blob/main/Cargo.toml

Mina side: https://github.com/MinaProtocol/mina/pull/12602)

## Stability?

I did run into some weird mutex bug that I can't reproduce anymore (https://github.com/supranational/pasta-msm/issues/2) so I'm putting this behind a `gpu` feature.

## Setup

You need a [CUDA-compatible GPU](https://developer.nvidia.com/cuda-gpus) for this to work. If you don't, it'll try to fall back on an AVX CPU implementation which seems to be somewhat faster than arkworks in my case.

You also need to have `nvcc` in your path. On Ubuntu I had to downgrade to g++-10 and c++-10 to make it work (make sure that `cc`, `c++`, also point to `gcc-10` and `g++-10` using `update-alternatives`)

## Run the bench

The bench can be ran with:

```
$ cargo criterion -p o1-utils --bench msm --features gpu
```

it only runs on small URS size. It's expected that the GPU implementation only starts getting interesting around MSM of size 128.

There's a simple test that measures a larger URS (2^18) can be run by doing:

```
$ cargo test --release --package o1-utils --features gpu -- --ignored large_msm --nocapture
```

## ADX implementation

The test running without the GPU impl but the ADX impl shows a 10ms difference for me:

```
$ cargo test --release --package o1-utils --features gpu -- --ignored large_msm --nocapture
running MSM of size 262144
cpu msm took: 281.071544ms
gpu msm took: 271.999473ms
```

if I do 100 of them the gap increases, so perhaps this is good

```
running MSM of size 262144
100 cpu msm took: 43.320211253s
100 gpu msm took: 35.781777069s
```

## Sanity of integration

Here's a flamegraph, from my reading our usage is not spending much time doing non-useful things (like converting types).

![flamegraph](https://user-images.githubusercontent.com/1316043/216749422-99e9c92d-e037-4616-9e3d-3748fc475842.svg)

## GPU numbers

I'm getting really bad numbers for the GPU implementation. It's basically almost 5x slower than the non-GPU implementation. My guess is that it's because I'm running things in WSL and I misconfigured something.

It'd be good to have someone running linux with a CUDA-compatible GPU to perform some bench here : o