@duguorong009 2023-02-05
@mimoo 
I found the request on the twitter.(https://twitter.com/cryptodavidw/status/1621948382683213824).
Here is one result I could provide.
```


Gnuplot not found, using plotters backend
arkworks msm of length 20                                                                            
                        time:   [194.16 µs 209.35 µs 226.38 µs]

Benchmarking supra msm of length 20: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.0s, enable flat sampling, or reduce sample count to 70.
supra msm of length 20  time:   [994.78 µs 994.90 µs 995.04 µs]                                    

arkworks msm of length 128                                                                            
                        time:   [292.83 µs 298.58 µs 304.91 µs]

supra msm of length 128 time:   [327.28 µs 350.70 µs 374.59 µs]                                    

arkworks msm of length 200                                                                            
                        time:   [394.92 µs 401.60 µs 409.36 µs]

supra msm of length 200 time:   [531.23 µs 574.51 µs 620.00 µs]                                    

arkworks msm of length 400                                                                            
                        time:   [658.72 µs 686.92 µs 719.91 µs]

Benchmarking supra msm of length 400: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.7s, enable flat sampling, or reduce sample count to 60.
supra msm of length 400 time:   [1.0556 ms 1.0877 ms 1.1170 ms]                                     
```

Here is the GPU info.
```
+-----------------------------------------------------------------------------+
| NVIDIA-SMI 510.108.03   Driver Version: 510.108.03   CUDA Version: 11.6     |
|-------------------------------+----------------------+----------------------+
| GPU  Name        Persistence-M| Bus-Id        Disp.A | Volatile Uncorr. ECC |
| Fan  Temp  Perf  Pwr:Usage/Cap|         Memory-Usage | GPU-Util  Compute M. |
|                               |                      |               MIG M. |
|===============================+======================+======================|
|   0  NVIDIA GeForce ...  Off  | 00000000:01:00.0  On |                  N/A |
| N/A   51C    P8    16W /  N/A |     54MiB /  8192MiB |     19%      Default |
|                               |                      |                  N/A |
+-------------------------------+----------------------+----------------------+
                                                                               
+-----------------------------------------------------------------------------+
| Processes:                                                                  |
|  GPU   GI   CI        PID   Type   Process name                  GPU Memory |
|        ID   ID                                                   Usage      |
|=============================================================================|
|    0   N/A  N/A      2415      G   /usr/lib/xorg/Xorg                 53MiB |
+-----------------------------------------------------------------------------+
```