@mimoo 2023-02-05
thanks @duguorong009 , looks like your GPU was truncated but I see cuda so most likely cuda compatible :o indeed these don't seem like great benchmarks as well. I'm wondering what's going on there...