@mimoo 2023-02-06
Note to myself: the implementation actually falls back to ADX (non-GPU) for MSM of size < 2^16